Move class to other folder
